# apps/core/internal/interfaces/httpserver/api/models.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for DirectoryPickerPurpose.
const (
	ProjectLocation DirectoryPickerPurpose = "project_location"
)

// Valid indicates whether the value is a known member of the DirectoryPickerPurpose enum.
func (e DirectoryPickerPurpose) Valid() bool {
	switch e {
	case ProjectLocation:
		return true
	default:
		return false
	}
}

// Defines values for ErrorCode.
const (
	ErrorCodeDirectoryPickerUnavailable ErrorCode = "directory_picker_unavailable"
	ErrorCodeInternalError              ErrorCode = "internal_error"
	ErrorCodeProjectLocationConflict    ErrorCode = "project_location_conflict"
	ErrorCodeProjectNotFound            ErrorCode = "project_not_found"
	ErrorCodeValidationFailed           ErrorCode = "validation_failed"
)

// Valid indicates whether the value is a known member of the ErrorCode enum.
func (e ErrorCode) Valid() bool {
	switch e {
	case ErrorCodeDirectoryPickerUnavailable:
		return true
	case ErrorCodeInternalError:
		return true
	case ErrorCodeProjectLocationConflict:
		return true
	case ErrorCodeProjectNotFound:
		return true
	case ErrorCodeValidationFailed:
		return true
	default:
		return false
	}
}

// Defines values for ProjectStage.
const (
	Concept            ProjectStage = "concept"
	Development        ProjectStage = "development"
	ReleasePreparation ProjectStage = "release_preparation"
	Released           ProjectStage = "released"
)

// Valid indicates whether the value is a known member of the ProjectStage enum.
func (e ProjectStage) Valid() bool {
	switch e {
	case Concept:
		return true
	case Development:
		return true
	case ReleasePreparation:
		return true
	case Released:
		return true
	default:
		return false
	}
}

// Defines values for ProjectStatus.
const (
	Active   ProjectStatus = "active"
	Archived ProjectStatus = "archived"
)

// Valid indicates whether the value is a known member of the ProjectStatus enum.
func (e ProjectStatus) Valid() bool {
	switch e {
	case Active:
		return true
	case Archived:
		return true
	default:
		return false
	}
}

// CreateProjectRequest defines model for CreateProjectRequest.
type CreateProjectRequest struct {
	Description *string `json:"description,omitempty"`
	Language    string  `json:"language"`

	// Location Existing absolute directory on the Core host.
	Location   string       `json:"location"`
	Name       string       `json:"name"`
	Stage      ProjectStage `json:"stage"`
	SteamAppId *int64       `json:"steam_app_id,omitempty"`
}

// DirectoryPickerPurpose defines model for DirectoryPickerPurpose.
type DirectoryPickerPurpose string

// Error defines model for Error.
type Error struct {
	Code        ErrorCode `json:"code"`
	Message     string    `json:"message"`
	Recoverable bool      `json:"recoverable"`
}

// ErrorCode defines model for Error.Code.
type ErrorCode string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error Error `json:"error"`
}

// Project defines model for Project.
type Project struct {
	CreatedAt   time.Time          `json:"created_at"`
	Description string             `json:"description"`
	Id          openapi_types.UUID `json:"id"`
	Language    string             `json:"language"`
	Location    string             `json:"location"`
	Name        string             `json:"name"`
	Stage       ProjectStage       `json:"stage"`
	Status      ProjectStatus      `json:"status"`
	SteamAppId  *int64             `json:"steam_app_id"`
	UpdatedAt   time.Time          `json:"updated_at"`
}

// ProjectList defines model for ProjectList.
type ProjectList struct {
	Items []Project `json:"items"`
}

// ProjectStage defines model for ProjectStage.
type ProjectStage string

// ProjectStatus defines model for ProjectStatus.
type ProjectStatus string

// SelectDirectoryRequest defines model for SelectDirectoryRequest.
type SelectDirectoryRequest struct {
	Purpose DirectoryPickerPurpose `json:"purpose"`
}

// SelectDirectoryResponse defines model for SelectDirectoryResponse.
type SelectDirectoryResponse struct {
	// Path Absolute directory path on the Core host, or null when cancelled.
	Path     *string `json:"path"`
	Selected bool    `json:"selected"`
}

// DirectoryPickerUnavailable defines model for DirectoryPickerUnavailable.
type DirectoryPickerUnavailable = ErrorResponse

// InternalError defines model for InternalError.
type InternalError = ErrorResponse

// ProjectLocationConflict defines model for ProjectLocationConflict.
type ProjectLocationConflict = ErrorResponse

// ProjectNotFound defines model for ProjectNotFound.
type ProjectNotFound = ErrorResponse

// ValidationError defines model for ValidationError.
type ValidationError = ErrorResponse

// CreateProjectJSONRequestBody defines body for CreateProject for application/json ContentType.
type CreateProjectJSONRequestBody = CreateProjectRequest

// SelectDirectoryJSONRequestBody defines body for SelectDirectory for application/json ContentType.
type SelectDirectoryJSONRequestBody = SelectDirectoryRequest