# packages/api-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: `${string}://${string}` | (string & {});
};

export type ProjectStage = 'concept' | 'development' | 'release_preparation' | 'released';

export type ProjectStatus = 'active' | 'archived';

export type DirectoryPickerPurpose = 'project_location';

export type SelectDirectoryRequest = {
  purpose: DirectoryPickerPurpose;
};

export type SelectDirectoryResponse = {
  selected: boolean;
  /**
   * Absolute directory path on the Core host, or null when cancelled.
   */
  path: string | null;
};

export type CreateProjectRequest = {
  name: string;
  description?: string;
  /**
   * Existing absolute directory on the Core host.
   */
  location: string;
  steam_app_id?: number;
  language: string;
  stage: ProjectStage;
};

export type Project = {
  id: string;
  name: string;
  description: string;
  location: string;
  steam_app_id: number | null;
  language: string;
  stage: ProjectStage;
  status: ProjectStatus;
  created_at: string;
  updated_at: string;
};

export type ProjectList = {
  items: Array<Project>;
};

export type Error = {
  code:
    | 'validation_failed'
    | 'project_not_found'
    | 'project_location_conflict'
    | 'directory_picker_unavailable'
    | 'internal_error';
  message: string;
  recoverable: boolean;
};

export type ErrorResponse = {
  error: Error;
};

export type SelectDirectoryData = {
  body: SelectDirectoryRequest;
  path?: never;
  query?: never;
  url: '/api/v1/system/select-directory';
};

export type SelectDirectoryErrors = {
  /**
   * Request validation failed.
   */
  400: ErrorResponse;
  /**
   * Core could not complete the request.
   */
  500: ErrorResponse;
  /**
   * The Core host cannot open a native directory picker.
   */
  503: ErrorResponse;
};

export type SelectDirectoryError = SelectDirectoryErrors[keyof SelectDirectoryErrors];

export type SelectDirectoryResponses = {
  /**
   * Directory selection completed or was cancelled by the user.
   */
  200: SelectDirectoryResponse;
};

export type SelectDirectoryResponse2 = SelectDirectoryResponses[keyof SelectDirectoryResponses];

export type ListProjectsData = {
  body?: never;
  path?: never;
  query?: never;
  url: '/api/v1/projects';
};

export type ListProjectsErrors = {
  /**
   * Core could not complete the request.
   */
  500: ErrorResponse;
};

export type ListProjectsError = ListProjectsErrors[keyof ListProjectsErrors];

export type ListProjectsResponses = {
  /**
   * Projects ordered from newest to oldest.
   */
  200: ProjectList;
};

export type ListProjectsResponse = ListProjectsResponses[keyof ListProjectsResponses];

export type CreateProjectData = {
  body: CreateProjectRequest;
  path?: never;
  query?: never;
  url: '/api/v1/projects';
};

export type CreateProjectErrors = {
  /**
   * Request validation failed.
   */
  400: ErrorResponse;
  /**
   * Another Project already references the same directory.
   */
  409: ErrorResponse;
  /**
   * Core could not complete the request.
   */
  500: ErrorResponse;
};

export type CreateProjectError = CreateProjectErrors[keyof CreateProjectErrors];

export type CreateProjectResponses = {
  /**
   * Project created.
   */
  201: Project;
};

export type CreateProjectResponse = CreateProjectResponses[keyof CreateProjectResponses];

export type GetProjectData = {
  body?: never;
  path: {
    /**
     * UUIDv7 Project identifier.
     */
    project_id: string;
  };
  query?: never;
  url: '/api/v1/projects/{project_id}';
};

export type GetProjectErrors = {
  /**
   * Request validation failed.
   */
  400: ErrorResponse;
  /**
   * Project does not exist.
   */
  404: ErrorResponse;
  /**
   * Core could not complete the request.
   */
  500: ErrorResponse;
};

export type GetProjectError = GetProjectErrors[keyof GetProjectErrors];

export type GetProjectResponses = {
  /**
   * Project detail.
   */
  200: Project;
};

export type GetProjectResponse = GetProjectResponses[keyof GetProjectResponses];